Refactor code to avoid duplication
