+++ app/backend/src/main/java/com/microsoft/openai/samples/rag/ask/approaches/RetrieveThenReadApproach.java
@@ -2,9 +2,6 @@ package com.microsoft.openai.samples.rag.ask.approaches;
 
 import com.azure.ai.openai.models.Completions;
 import com.azure.ai.openai.models.CompletionsOptions;
-import com.azure.core.util.Context;
-import com.azure.search.documents.SearchDocument;
-import com.azure.search.documents.models.*;
 import com.azure.search.documents.util.SearchPagedIterable;
 import com.microsoft.openai.samples.rag.approaches.ContentSource;
 import com.microsoft.openai.samples.rag.approaches.RAGApproach;
@@ -16,7 +13,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 /**
  * Simple retrieve-then-read implementation, using the Cognitive Search and OpenAI APIs directly. It first retrieves
@@ -70,72 +69,14 @@ public class RetrieveThenReadApproach implements RAGApproach<String, RAGResponse
 
     private  CompletionsOptions buildCompletionsOptions(SemanticSearchAskPrompt retrieveThenReadPrompt) {
         CompletionsOptions completionsOptions = new CompletionsOptions(new ArrayList<>(Collections.singletonList(retrieveThenReadPrompt.getFormattedPrompt())));
-
         // Due to a potential bug when using JVM 17 and java openai SDK 1.0.0-beta.2, we need to provide default for all properties to avoid 404 bad Request on the server
-        completionsOptions.setMaxTokens(1024);
-        completionsOptions.setTemperature(0.3);
         completionsOptions.setStop(List.of("\n"));
-        completionsOptions.setLogitBias(new HashMap<>());
-        completionsOptions.setEcho(false);
-        completionsOptions.setN(1);
-        completionsOptions.setStream(false);
-        completionsOptions.setUser( "search-openai-demo-java");
-        completionsOptions.setPresencePenalty(0.0);
-        completionsOptions.setFrequencyPenalty(0.0);
-        completionsOptions.setBestOf(1);
-
-
-        return completionsOptions;
+        return fillCommonCompletionsOptions(completionsOptions);
     }
 
-    private SearchPagedIterable getCognitiveSearchResults(String question, RAGOptions options) {
-        var searchOptions = new SearchOptions();
-
-        Optional.ofNullable(options.getTop()).ifPresentOrElse(
-                searchOptions::setTop,
-                () -> searchOptions.setTop(3));
-        Optional.ofNullable(options.getExcludeCategory())
-                .ifPresentOrElse(
-                        value -> searchOptions.setFilter("category ne '%s'".formatted(value.replace("'", "''"))),
-                        () -> searchOptions.setFilter(null));
-
-        Optional.ofNullable(options.isSemanticRanker()).ifPresent(isSemanticRanker -> {
-           if(isSemanticRanker) {
-               searchOptions.setQueryType(QueryType.SEMANTIC);
-               searchOptions.setQueryLanguage(QueryLanguage.EN_US);
-               searchOptions.setSpeller(QuerySpellerType.LEXICON);
-               searchOptions.setSemanticConfigurationName("default");
-               searchOptions.setQueryCaption(QueryCaptionType.EXTRACTIVE);
-               searchOptions.setQueryCaptionHighlightEnabled(false);
-           }
-        });
-
-        return this.cognitiveSearchProxy.search(question, searchOptions, Context.NONE);
-    }
-
-    private List<ContentSource> buildSourcesFromSearchResults(RAGOptions options, SearchPagedIterable searchResults) {
-        List<ContentSource> sources = new ArrayList<>();
-
-        searchResults.iterator().forEachRemaining(result ->
-        {
-           var searchDocument = result.getDocument(SearchDocument.class);
-
-           /*
-            If captions is enabled the content source is taken from the captions generated by the semantic ranker.
-            Captions are appended sequentially and separated by a dot.
-            */
-           if(options.isSemanticCaptions()) {
-               StringBuilder sourcesContentBuffer = new StringBuilder();
-
-               result.getCaptions().forEach(caption -> sourcesContentBuffer.append(caption.getText()).append("."));
-
-               sources.add(new ContentSource((String)searchDocument.get("sourcepage"), sourcesContentBuffer.toString()));
-           } else {
-               //If captions is disabled the content source is taken from the cognitive search index field "content"
-               sources.add(new ContentSource((String) searchDocument.get("sourcepage"), (String) searchDocument.get("content")));
-           }
-        });
-        return sources;
+    @Override
+    public CognitiveSearchProxy getCognitiveSearchProxy() {
+        return this.cognitiveSearchProxy;
     }
 
 }
