+++ app/backend/src/main/java/com/microsoft/openai/samples/rag/chat/approaches/ChatReadRetrieveReadApproach.java
@@ -1,9 +1,6 @@
 package com.microsoft.openai.samples.rag.chat.approaches;
 
 import com.azure.ai.openai.models.*;
-import com.azure.core.util.Context;
-import com.azure.search.documents.SearchDocument;
-import com.azure.search.documents.models.*;
 import com.azure.search.documents.util.SearchPagedIterable;
 import com.microsoft.openai.samples.rag.approaches.ContentSource;
 import com.microsoft.openai.samples.rag.approaches.RAGApproach;
@@ -88,21 +85,9 @@ public class ChatReadRetrieveReadApproach implements RAGApproach<ChatGPTConversa
 
     private CompletionsOptions buildCompletionsOptions(CognitiveSearchQueryPrompt retrieveThenReadPrompt) {
         CompletionsOptions completionsOptions = new CompletionsOptions(new ArrayList<>(Collections.singletonList(retrieveThenReadPrompt.getFormattedPrompt())));
-
         // Due to a potential bug in using JVM 17 and java open SDK 1.0.0-beta.2, we need to provide default for all properties to avoid 404 bad Request on the server
-        completionsOptions.setMaxTokens(1024);
-        completionsOptions.setTemperature(0.3);
         completionsOptions.setStop(new ArrayList<>( Arrays.asList("<|im_end|>","<|im_start|>")));
-        completionsOptions.setLogitBias(new HashMap<>());
-        completionsOptions.setEcho(false);
-        completionsOptions.setN(1);
-        completionsOptions.setStream(false);
-        completionsOptions.setUser( "search-openai-demo-java");
-        completionsOptions.setPresencePenalty(0.0);
-        completionsOptions.setFrequencyPenalty(0.0);
-        completionsOptions.setBestOf(1);
-
-        return completionsOptions;
+        return fillCommonCompletionsOptions(completionsOptions);
     }
 
     private ChatCompletionsOptions buildChatCompletionsOptions(SemanticSearchChat semanticSearchChat) {
@@ -115,61 +100,16 @@ public class ChatReadRetrieveReadApproach implements RAGApproach<ChatGPTConversa
         completionsOptions.setLogitBias(new HashMap<>());
         completionsOptions.setN(1);
         completionsOptions.setStream(false);
-        completionsOptions.setUser( "search-openai-demo-java");
+        completionsOptions.setUser("search-openai-demo-java");
         completionsOptions.setPresencePenalty(0.0);
         completionsOptions.setFrequencyPenalty(0.0);
-
 
         return completionsOptions;
     }
 
-    private SearchPagedIterable getCognitiveSearchResults(String question, RAGOptions options) {
-        var searchOptions = new SearchOptions();
-
-        Optional.ofNullable(options.getTop()).ifPresentOrElse(
-                searchOptions::setTop,
-                () -> searchOptions.setTop(3));
-        Optional.ofNullable(options.getExcludeCategory())
-                .ifPresentOrElse(
-                        value -> searchOptions.setFilter("category ne '%s'".formatted(value.replace("'", "''"))),
-                        () -> searchOptions.setFilter(null));
-
-        Optional.ofNullable(options.isSemanticRanker()).ifPresent(isSemanticRanker -> {
-            if(isSemanticRanker) {
-                searchOptions.setQueryType(QueryType.SEMANTIC);
-                searchOptions.setQueryLanguage(QueryLanguage.EN_US);
-                searchOptions.setSpeller(QuerySpellerType.LEXICON);
-                searchOptions.setSemanticConfigurationName("default");
-                searchOptions.setQueryCaption(QueryCaptionType.EXTRACTIVE);
-                searchOptions.setQueryCaptionHighlightEnabled(false);
-            }
-        });
-
-        return this.cognitiveSearchProxy.search(question, searchOptions, Context.NONE);
+    @Override
+    public CognitiveSearchProxy getCognitiveSearchProxy() {
+        return this.cognitiveSearchProxy;
     }
 
-    private List<ContentSource> buildSourcesFromSearchResults(RAGOptions options, SearchPagedIterable searchResults) {
-        List<ContentSource> sources = new ArrayList<>();
-
-        searchResults.iterator().forEachRemaining(result ->
-        {
-            var searchDocument = result.getDocument(SearchDocument.class);
-
-            /*
-             If captions is enabled the content source is taken from the captions generated by the semantic ranker.
-             Captions are appended sequentially and separated by a dot.
-             */
-            if(options.isSemanticCaptions()) {
-                StringBuilder sourcesContentBuffer = new StringBuilder();
-
-                result.getCaptions().forEach(caption -> sourcesContentBuffer.append(caption.getText()).append("."));
-
-                sources.add(new ContentSource((String)searchDocument.get("sourcepage"), sourcesContentBuffer.toString()));
-            } else {
-                //If captions is disabled the content source is taken from the cognitive search index field "content"
-                sources.add(new ContentSource((String) searchDocument.get("sourcepage"), (String) searchDocument.get("content")));
-            }
-        });
-        return sources;
-    }
 }
